Criados os testes para o repositório de professores

+++ sigai/app/Repositories/Eloquent/ProfessorRepository.php
@@ -4,6 +4,7 @@ use App\Models\Curso;
 use App\Models\Professor;
 use App\Models\User;
 
+use App\Repositories\Contracts\CursoRepositoryContract;
 use App\Repositories\Contracts\ProfessorRepositoryContract;
 
 use App\Exceptions\NotFoundError;
@@ -12,10 +13,11 @@ use App\Exceptions\ServerError;
 use \DB;
 use \Lang;
 use \Hash;
+use \Log;
 
 class ProfessorRepository extends BaseRepository implements ProfessorRepositoryContract
 {
-    public static function findById($id)
+    public function findById($id)
     {
         $professor = Professor::join('usuarios', 'professores.id', '=', 'usuarios.id')
                               ->where('usuarios.id', $id)->first();
@@ -27,7 +29,7 @@ class ProfessorRepository extends BaseRepository implements ProfessorRepositoryC
         return $professor;
     }
     
-    public static function findByIdWith($id, array $relations)
+    public function findByIdWith($id, array $relations)
     {
         $professor = Professor::join('usuarios', 'professores.id', '=', 'usuarios.id')
                               ->where('usuarios.id', $id)
@@ -40,7 +42,7 @@ class ProfessorRepository extends BaseRepository implements ProfessorRepositoryC
         return $professor;
     }
     
-    public static function findByMatricula($matricula)
+    public function findByMatricula($matricula)
     {
         $professor = Professor::join('usuarios', 'professores.id', '=', 'usuarios.id')
                               ->where('usuarios.matricula', $matricula)->first();
@@ -52,7 +54,7 @@ class ProfessorRepository extends BaseRepository implements ProfessorRepositoryC
         return $professor;
     }
     
-    public static function findByNome($nome)
+    public function findByNome($nome)
     {
         $professor = Professor::join('usuarios', 'professores.id', '=', 'usuarios.id')
                               ->where('usuarios.nome', $nome)->first();
@@ -64,15 +66,15 @@ class ProfessorRepository extends BaseRepository implements ProfessorRepositoryC
         return $professor;
     }
     
-    public static function searchByNameAndMatricula($query = null, $turmaId = null)
+    public function searchByNameAndMatricula($query = null, $turmaId = null)
     {
         $q = Professor::join('usuarios', 'professores.id', '=', 'usuarios.id');
         
         if (!empty($turmaId)) {
-            $q->whereNotIn('professores.id', function($query) {
+            $q->whereNotIn('professores.id', function($query) use ($turmaId) {
                 $query->select('professor_id')
                       ->from('professores_turmas')
-                      ->where('turma_id', 2);
+                      ->where('turma_id', $turmaId);
             });
         }
                         
@@ -84,15 +86,15 @@ class ProfessorRepository extends BaseRepository implements ProfessorRepositoryC
         return $q->get();
     }
     
-    public static function paginate($orderBy = 'usuarios.nome', $perPage = 10)
+    public function paginate($orderBy = 'usuarios.nome', $perPage = 10)
     {
         $alunos = Professor::with('cursoOrigem')
                            ->join('usuarios', 'professores.id', '=', 'usuarios.id')
 	                       ->orderBy($orderBy)->paginate($perPage);
 	    return $alunos;
     }
-    
-    public static function insert(array $data)
+
+    public function insert(array $data)
     {
         $usuario = new User;
         $usuario->matricula = $data['matricula'];
@@ -104,29 +106,30 @@ class ProfessorRepository extends BaseRepository implements ProfessorRepositoryC
             throw new ServerError(Lang::get('professores.create_error'));
         }
        
-        $professor = new Professor();
+        $professor = new Professor;
         $professor->id = $usuario->id;
-        
-        $curso = CursoRepository::findById($data['curso_origem_id']);
-        $professor->cursoOrigem()->associate($curso);
-       
+
+        if (isset($data['curso_origem'])) {
+            $professor->cursoOrigem()->associate($data['curso_origem']);
+        }
+
         if (!$professor->save()) {
             throw new ServerError(Lang::get('professores.create_error'));
         }
         
-        $professor->cursoOrigem = $curso;
         $professor->usuario = $usuario;
+        $professor->id = $usuario->id;
         
         return $professor;
     }
     
-    public static function updateByMatricula(array $data, $matricula)
+    public function updateByMatricula(array $data, $matricula)
     {
         $professor = self::findByMatricula($matricula);
 
-        $professor->usuario->matricula = $data['matricula'];
-        $professor->usuario->nome      = $data['nome'];
-        $professor->usuario->email     = $data['email'];
+        $professor->usuario->matricula = array_get($data, 'matricula');
+        $professor->usuario->nome      = array_get($data, 'nome');
+        $professor->usuario->email     = array_get($data, 'email');
         
         if (isset($data['password']) && strlen($data['password']) > 0) {
             $professor->usuario->password = Hash::make($data['password']);
@@ -136,8 +139,9 @@ class ProfessorRepository extends BaseRepository implements ProfessorRepositoryC
             throw new ServerError(Lang::get('professores.create_error'));
         }
 
-        $professor->cursoOrigem()->associate(
-            CursoRepository::findById($data['curso_origem_id']));
+        if (isset($data['curso_origem'])) {
+            $professor->cursoOrigem()->associate($data['curso_origem']);
+        }
        
         if (!$professor->save()) {
             throw new ServerError(Lang::get('professores.create_error'));
@@ -146,20 +150,32 @@ class ProfessorRepository extends BaseRepository implements ProfessorRepositoryC
         return $professor;
     }
     
-    public static function deleteByMatricula($matricula)
+    public function deleteByMatricula($matricula)
     {
-        $professor = self::findByMatricula($matricula);
-        $usuario   = $professor->usuario;
+        $professor = Professor::join('usuarios', 'professores.id', '=', 'usuarios.id')
+            ->with('usuario.cursos')
+            ->where('usuarios.matricula', $matricula)->first();
+
+        if ($professor == null) {
+            throw new NotFoundError(Lang::get('professores.not_found'));
+        }
         
         DB::beginTransaction();
 	    
 	    try {
+            foreach ($professor->usuario->cursos as $curso) {
+                $curso->coordenador()->dissociate();
+                $curso->save();
+            }
+
 	        $professor->statusDiarios()->delete();
 	        $professor->turmas()->detach();
+
 	        $professor->delete();
-	        $usuario->delete();
+            $professor->usuario->delete();
 	    } catch (\Exception $e) {
 	        DB::rollback();
+            Log::error($e->getMessage(), ['context' => $e->getTraceAsString()]);
 	        throw new ServerError(Lang::get('professores.remove_error'));
 	    }
